+++ janitor/lib/mastodon_helper.py
@@ -21,7 +21,9 @@ class MastodonHelper:
 
     @staticmethod
     def get_instance(
-        config: Config, connection_params: MastodonConnectionParams = None
+        config: Config,
+        connection_params: MastodonConnectionParams = None,
+        base_path: str = None
     ) -> Mastodon:
         logger = logging.getLogger(config.get("logger.name"))
 
@@ -32,13 +34,18 @@ class MastodonHelper:
             connection_params = MastodonConnectionParams.from_dict(config.get("mastodon"))
 
         instance_type = MastodonHelper.valid_or_raise(connection_params.instance_type)
+        user_file = connection_params.credentials.user_file
+        client_file = connection_params.credentials.client_file
+        if base_path is not None:
+            user_file = os.path.join(base_path, user_file)
+            client_file = os.path.join(base_path, client_file)
 
         # All actions are done under a Mastodon API instance
         logger.info("Starting new Mastodon API instance")
-        if (os.path.exists(connection_params.credentials.user_file)):
+        if (os.path.exists(user_file)):
             logger.debug("Reusing stored User Credentials")
             mastodon = MastodonHelper.WRAPPER[instance_type](
-                access_token=connection_params.credentials.user_file,
+                access_token=user_file,
                 feature_set=MastodonHelper.FEATURE_SET_BY_INSTANCE_TYPE[instance_type]
                 if instance_type in [MastodonHelper.TYPE_MASTODON, MastodonHelper.TYPE_PLEROMA]
                 else None
@@ -46,7 +53,7 @@ class MastodonHelper:
         else:
             logger.debug("Using Client Credentials")
             mastodon = MastodonHelper.WRAPPER[instance_type](
-                client_id=connection_params.credentials.client_file,
+                client_id=client_file,
                 api_base_url=connection_params.api_base_url,
                 feature_set=MastodonHelper.FEATURE_SET_BY_INSTANCE_TYPE[instance_type]
                 if instance_type in [MastodonHelper.TYPE_MASTODON, MastodonHelper.TYPE_PLEROMA]
@@ -58,7 +65,7 @@ class MastodonHelper:
             mastodon.log_in(
                 connection_params.credentials.user.email,
                 connection_params.credentials.user.password,
-                to_file=connection_params.credentials.user_file
+                to_file=user_file
             )
 
         return mastodon