+++ tests/lib/test_publisher.py
@@ -20,6 +20,7 @@ CONFIG = {
 }
 
 _mocked_mastodon_instance: Mastodon = Mock()
+_mocked_queue_instance = Mock()
 
 
 def patched_config_init(self):
@@ -34,22 +35,31 @@ def patched_generic_init(self, config: Config):
     pass
 
 
-def get_instance() -> Publisher:
+def get_instance(config_instance: Config = None) -> Publisher:
     _mocked_mastodon_instance.__class__ = Mastodon
     _mocked_mastodon_instance.status_post = Mock()
     _mocked_mastodon_instance.status_post.return_value = {"id": 123}
     _mocked_mastodon_instance.media_post = Mock()
     _mocked_mastodon_instance.media_post.return_value = {"id": 456}
+    _mocked_queue_instance.return_value = None
 
     with patch.object(Config, "__init__", new=patched_config_init):
         with patch.object(Config, "get", new=patched_config_get):
-            with patch.object(Queue, "__init__", new=patched_generic_init):
+            with patch.object(Queue, "__init__", new=_mocked_queue_instance):
                 with patch.object(Formatter, "__init__", new=patched_generic_init):
-                    return Publisher(config=Config(), mastodon=_mocked_mastodon_instance)
+                    if config_instance is None:
+                        config_instance = Config()
+                    return Publisher(
+                        config=config_instance,
+                        mastodon=_mocked_mastodon_instance,
+                        base_path="bla"
+                    )
 
 
 def test_initialize():
-    publisher = get_instance()
+    config_instance = Config()
+
+    publisher = get_instance(config_instance)
 
     assert isinstance(publisher, Publisher)
     assert isinstance(publisher._config, Config)
@@ -57,6 +67,7 @@ def test_initialize():
     assert isinstance(publisher._queue, Queue)
     assert isinstance(publisher._formatter, Formatter)
     assert isinstance(publisher._mastodon, Mastodon)
+    _mocked_queue_instance.assert_called_once_with(config_instance, base_path="bla")
 
 
 @pytest.fixture