+++ CHANGELOG.md
@@ -6,11 +6,15 @@
 
 - Runner to monitor Git repositories and publish the identified new version from the Changelog
 - New `TODO.md` list ;-)
+- Added a main runner to serve as a Command Line Interface tool.
+- Added the possibility from the config file to run the listener in debug mode
+- Added a new `definitions` root module to serve global info like the root directory.
 
 ### Changed
 
 - `MastodonHelper` now is able to receive connection params injected
-- Bumped pyxavi to 0.5.1
+- Bumped pyxavi to 0.5.2
+- Moved all runners into an internal directory. They can't be called directly anymore.
 
 ## [v0.3.0](https://github.com/XaviArnaus/janitor/releases/tag/v0.3.0)